+++ sample-app/src/main/java/com/test/JaxrsApplication.java
@@ -1,5 +1,7 @@
 package com.test;
 
+import org.springframework.stereotype.Component;
+
 import javax.ws.rs.ApplicationPath;
 import javax.ws.rs.core.Application;
 
@@ -8,6 +10,7 @@ import javax.ws.rs.core.Application;
  *
  * Created by facarvalho on 12/7/15.
  */
+@Component
 @ApplicationPath("/sample-app/")
 public class JaxrsApplication extends Application {
 }